requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2024.2
requests==2.31.0